### 2016-05-09 #16 @whedon
Submission: R3D2: Relativistic Reactive Riemann problem solver for Deflagrations and Detonations
**Submitting author:** @harpolea (<a href="http://orcid.org/0000-0002-1530-781X">Alice Harpole</a>)
**Repository:** <a href="https://github.com/harpolea/r3d2" target ="_blank">https://github.com/harpolea/r3d2</a>
**Version:** v1.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.51096" target ="_blank">http://dx.doi.org/10.5281/zenodo.51096</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51096">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00016.pdf](https://github.com/openjournals/joss-reviews/files/286440/10.21105.joss.00016.pdf)
