### 2016-05-13 d31cb9c
Minor changes
