### 2016-05-13 10dacf8
Code for finding critical tangential velocity for varying rho
