+++ investigate_wave_pattern.py
@@ -15,6 +15,8 @@ def check_wave_pattern(U_l, U_r, vt_side, vts=[-0.9,0.5,0.0,-0.5,0.9]):
     """
     flat_patterns = make_flat_patterns(U_l, U_r, vts, vt_side)
 
+    #print(flat_patterns)
+
     # generate list of pairs
     pairs = list(combinations(range(len(vts)), 2))
 
@@ -46,14 +48,18 @@ def make_flat_patterns(U_l, U_r, vts, vt_side):
         eos_r = U_r.eos
 
     for vt in vts:
+        #print('vt = {}'.format(vt))
         # first change the vt
         if vt_side == 'l':
             U_l = State(rho_l, v_l, vt, eps_l, eos_l)
-            U_r = State(rho_r, v_r, 0.0, eps_r, eos_r)
+            #U_r = State(rho_r, v_r, 0.0, eps_r, eos_r)
         else:
             U_r = State(rho_r, v_r, vt, eps_r, eos_r)
-        rp = RiemannProblem(U_l, U_r)
-        wave_patterns.append(rp.waves)
+        try:
+            rp = RiemannProblem(U_l, U_r)
+            wave_patterns.append(rp.waves)
+        except:
+            break
 
     # now check if all the wave patterns are the same
 
@@ -76,7 +82,7 @@ def find_critical_vt(U_l, U_r, vt_side):
     the initial pass, then a root may be missed if the wave pattern changes
     to a different pattern then back again.
     """
-    vts = np.linspace(0., 0.9, num=100)
+    vts = np.linspace(0., 0.999, num=100)
     tolerance = 1.e-4
 
     def bisect(vt0, vtend, tol=tolerance):
@@ -92,7 +98,7 @@ def find_critical_vt(U_l, U_r, vt_side):
             vthalf = 0.5 * (vt0 + vtend)
             flat_patterns = make_flat_patterns(U_l, U_r, [vt0, vthalf], vt_side)
 
-            if not flat_patterns[0] == flat_patterns[1]:
+            if len(flat_patterns) > 1 and not flat_patterns[0] == flat_patterns[1]:
                 vtend = vthalf
             else:
                 vt0 = vthalf
@@ -110,7 +116,7 @@ def find_critical_vt(U_l, U_r, vt_side):
     critical_vts = []
     critical_patterns = []
 
-    for i in range(len(vts) - 1):
+    for i in range(len(flat_patterns) - 2):
         if not flat_patterns[i] == flat_patterns[i+1]:
             vt, pattern = bisect(vts[i], vts[i+1])
             critical_vts.append(vt)
@@ -123,13 +129,46 @@ def find_critical_vt(U_l, U_r, vt_side):
     for i, v in enumerate(critical_vts):
         print('vt: {}, patterns: {} -> {}'.format(v, ', '.join(critical_patterns[i][0]), ', '.join(critical_patterns[i][1])))
 
+    return critical_vts
+
+def vary_rho(U_l, U_r, vt_side, rhos):
+
+    critical_vts = np.zeros_like(rhos)
+
+    if vt_side == 'l':
+        v_l = U_l.v
+        eps_l = U_l.eps
+        eos_l = U_l.eos
+    else:
+        v_r = U_r.v
+        eps_r = U_r.eps
+        eos_r = U_r.eos
+
+    for i, rho in enumerate(rhos):
+
+        # first change the rho
+        if vt_side == 'l':
+            U_l = State(rho, v_l, 0.0, eps_l, eos_l)
[…]
