### 2016-05-13 62df343
Added plotting
