### 2016-05-18 #16 @kyleniemeyer
Haven't forgotten about this, just been busy... plan to review this in the next day or so.
