### 2016-05-23 #16 @kyleniemeyer
OK, finally performed my review (sorry for the delay!). Here are my comments:

* No performance claims (other than in one place stating performance is poor), so technically this is question is satisfied.

* Typo in Usage section of the README (also first page of the docs): `eos = eos_defn.eos_gamma_law(5.0/3.0)` should be `eos = eos_defns.eos_gamma_law(5.0/3.0)`. In addition, that section could be improved with a bit more detail about what can be done with the `rp` object once the Riemann Problem has been solved. I can see quite nice detailed examples in the documentation, but something to complete that usage section (what to do with that result, for example) might be good.

* The code itself is not documented, and the online documentation (e.g., "Indices and Tables" section it empty/leads to broken pages).

* Regarding the statement of need, the paper (and software README) give fairly technical descriptions, and could both use a higher-level description of need/intended audience that would be more easily understood by a general reader. For example, many technical readers may not immediately know what the Riemann problem is, so a gentler introduction would help.

(I'll add that the detailed examples given in the [documentation on readthedocs](http://r3d2.readthedocs.io/) are great!)
