### 2016-05-23 96dbda2
fixed typo eos_defn -> eos_defns
Part of JOSS review in open journals/joss-reviews#16
