### 2016-05-23 #16 @kyleniemeyer
In fact, I went ahead and corrected the typo in the docs directly via harpolea/r3d2#2
