### 2016-05-23 #16 @kyleniemeyer
Of course, any other eyes on this (e.g., @gravypod) can only help.
