### 2016-05-23 d072830
Fix typo following PR #2 (eos_defn -> eos_defns).
