### 2016-05-23 2d1cd88
Document how to examine the RiemannProblem object.
