### 2016-05-23 fb93ab3
Extend the description of the Riemann Problem.
Explain the point of the RP briefly.
