### 2016-05-23 e6365db
Start improving documentation.
