### 2016-05-23 e38d006
Improve documentation in all files.
Closer match to numpydoc standards; see
http://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_numpy.html
