+++ r3d2/wave.py
@@ -3,6 +3,10 @@
 Created on Mon Feb 15 11:28:53 2016
 
 @author: ih3
+
+Wave class.
+
+Core functions and classes that solve across a wave, or wave section.
 """
 
 import numpy
@@ -105,6 +109,33 @@ def mass_flux_squared(q_start, p_end, unknown_eos=None):
     return j2, rho, eps, dp
 
 def deflagration_root(p_0_star, q_precursor, unknown_eos, wavenumber, label):
+    """
+    Find the CJ deflagration.
+    
+    Find the limiting case between stable and unstable deflagrations. Can also
+    be used for the detonation case.
+    
+    Parameters
+    ----------
+    
+    p_0_star : double
+        Pressure to match to.
+    q_precursor : State
+        Known State
+    unknown_eos : EOS
+        Equation of State after the reaction has taken place
+    wavenumber : int
+        Indicates if this is the left (0), central (1) or right (2) wave
+    label : string
+        Optional label of the resulting State
+        
+    Returns
+    -------
+    
+    residual : double
+        Residual function to be minimized
+    """
+    
     lr_sign = wavenumber - 1
     j2, rho, eps, dp = mass_flux_squared(q_precursor, p_0_star, unknown_eos)
     if j2 < 0:
@@ -131,6 +162,29 @@ def deflagration_root(p_0_star, q_precursor, unknown_eos, wavenumber, label):
     return q_unknown.wavespeed(wavenumber) - v_deflagration
 
 def precursor_root(p_0_star, q_known, wavenumber):
+    """
+    Find the precursor shock.
+    
+    For a detonation, the temperature needs to be raised across a shock for the
+    reaction to take place.
+    
+    Parameters
+    ----------
+    
+    p_0_star : double
+        Pressure to match to.
+    q_known : State
+        Known State
+    wavenumber : int
+        Indicates if this is the left (0), central (1) or right (2) wave
+        
+    Returns
+    -------
+    
+    residual : double
+        Residual function to be minimized
+    """
+    
     shock = Shock(q_known, p_0_star, wavenumber)
     q_precursor = shock.q_end
     t_precursor = q_precursor.eos['t_from_rho_eps'](
@@ -152,6 +206,41 @@ def precursor_root(p_0_star, q_known, wavenumber):
 
 def post_discontinuity_state(p_star, q_start, lr_sign, label, j2, rho, eps, dp,
                              eos_end = None):
+    """
+    Give the state across a discontinuity.
+    
+    This code is common to all discontinuities.
+    
+    Parameters
+    ----------
+    
+    p_star : double
+        Post-discontinuity pressure
+    q_start : State
+        Known State
+    lr_sign : int
+        -1 for a left going wave, +1 for a right going wave
+    label : string
+        Optional label of the post-shock State
+    j2 : double
+        Square of the mass flux across the wave
+    rho : double
+        Post-discontinuity density 
+    eps : double
+        Post-discontinuity specific internal energy
+    dp : double
+        Jump in pressure across the discontinuity
+    eos_end : EOS
+        Equation of State on the other side of the discontinuity, if different
+    
+    Returns
+    -------
+    
+    v_shock : double
+        Shock speed
+    q_end : State
+        State on the other side of the discontinuity.
+    """
     if eos_end is None:
         eos_end = q_start.eos
     j = numpy.sqrt(j2)
@@ -169,8 +258,33 @@ def post_discontinuity_state(p_star, q_start, lr_sign, label, j2, rho, eps, dp,
 
 class WaveSection(object):
     """
-    A wave section is a part of a wave. It has a single type, such as a shock
-    or rarefaction.
+    A wave section is a single type of solution where the State varies.
+
+    Parameters
[…]
