### 2016-05-24 6aa957f
Docstrings for utils functions
