+++ r3d2/utils.py
@@ -12,7 +12,22 @@ from matplotlib.transforms import offset_copy
 import matplotlib.transforms as mtrans
 from . import wave
 
-def rarefaction(ps, q_start, lr_sign):
+def _rarefaction(ps, q_start, lr_sign):
+    """
+    Finds the velocities across a rarefaction wave given the pressures, known
+    state on one side of the wave and direction the wave is travelling.
+
+    Parameters
+    ----------
+
+    ps : array
+        Array of pressures
+    q_start : State object
+        Known initial state on one side of the wave.
+    lr_sign : integer
+        Direction in which the wave is travelling: -1 if left-going, +1 if
+        right-going
+    """
     vs = numpy.zeros_like(ps)
     for i, p in enumerate(ps):
         w_all = odeint(wave.rarefaction_dwdp,
@@ -24,7 +39,23 @@ def rarefaction(ps, q_start, lr_sign):
         vs[i] = w_all[-1, 1]
     return vs
 
-def shock(ps, q_start, lr_sign):
+def _shock(ps, q_start, lr_sign):
+    """
+    Finds the velocities across a shock wave given the pressures, known
+    state on one side of the wave and direction the wave is travelling.
+
+    Parameters
+    ----------
+
+    ps : array
+        Array of pressures
+    q_start : State object
+        Known initial state on one side of the wave.
+    lr_sign : integer
+        Direction in which the wave is travelling: -1 if left-going, +1 if
+        right-going
+    """
+
     vs = numpy.zeros_like(ps)
     for i, p in enumerate(ps):
         j2, rho, eps, diffp = wave.mass_flux_squared(q_start,
@@ -36,7 +67,23 @@ def shock(ps, q_start, lr_sign):
         vs[i] = q_end.v
     return vs
 
-def deflagration(ps, q_start, lr_sign):
+def _deflagration(ps, q_start, lr_sign):
+    """
+    Finds the velocities across a deflagration wave given the pressures, known
+    state on one side of the wave and direction the wave is travelling.
+
+    Parameters
+    ----------
+
+    ps : array
+        Array of pressures
+    q_start : State object
+        Known initial state on one side of the wave.
+    lr_sign : integer
+        Direction in which the wave is travelling: -1 if left-going, +1 if
+        right-going
+    """
+
     vs = numpy.zeros_like(ps)
     for i, p in enumerate(ps):
         j2, rho, eps, diffp = wave.mass_flux_squared(q_start,
@@ -59,7 +106,23 @@ def deflagration(ps, q_start, lr_sign):
         vs[i] = q_end.v
     return vs
 
-def detonation(ps, q_start, lr_sign):
+def _detonation(ps, q_start, lr_sign):
+    """
+    Finds the velocities across a detonation wave given the pressures, known
+    state on one side of the wave and direction the wave is travelling.
+
+    Parameters
+    ----------
+
+    ps : array
+        Array of pressures
+    q_start : State object
+        Known initial state on one side of the wave.
+    lr_sign : integer
+        Direction in which the wave is travelling: -1 if left-going, +1 if
+        right-going
+    """
+
     vs = numpy.zeros_like(ps)
     for i, p in enumerate(ps):
[…]
