### 2016-05-24 #16 @harpolea
Thanks @kyleniemeyer - we've improved the documentation (adding docstrings to functions and removing the redundant 'Indices and Tables' section from the online docs) and the statement of need: harpolea/r3d2#3, harpolea/r3d2#4, harpolea/r3d2#5
