### 2016-05-27 #16 @kyleniemeyer
@harpolea those changes look good—did you update the paper with the new statement of need?

@arfon: where would an updated paper show up?
