### 2016-05-27 #16 @arfon
> @arfon: where would an updated paper show up?

I updated the paper in the original review issue.
