### 2016-05-27 #16 @harpolea
@kyleniemeyer Yes, we updated the paper to include more background information on the Riemann problem and its uses
