### 2016-05-27 #16 @kyleniemeyer
OK, at this point I think all the review criteria are satisfied.

@arfon recommend accepting now! (meta ?: should I let you add the "accepted" label, or do that myself?)
