### 2016-05-27 #16 @arfon
> @arfon recommend accepting now! (meta ?: should I let you add the "accepted" label, or do that myself?)

:boom: 🎉
