### 2016-05-27 #16 @arfon
@harpolea: http://dx.doi.org/10.21105/joss.00016 :zap: :boom: :rocket:
